test(redirect-to-project-create): do not use beforeEach
